Collect formatting tests in `Formatting` module
